LAB TASK 12/task12/requirements.txt
Flask==2.3.3
sentence-transformers==5.4.1
faiss-cpu==1.7.4
numpy==1.24.3